==> exploration/01_webscraping.ipynb <==
import requests
import pandas as pd
from bs4 import BeautifulSoup
from geopy.geocoders import Nominatim, GoogleV3
import numpy as np
# from googlemaps import GoogleMaps
import time
import folium
import pickle

url = r'https://wars.vote4.hk/en/high-risk'
hk_latitude, hk_longitude = 22.2793278, 114.1628131
page = requests.get(url)
soup = BeautifulSoup(page.content, 'html.parser')

html_addresses = soup.find_all(
    'span', 
    class_=r'MuiTypography-root MuiTypography-h6 MuiTypography-colorTextPrimary'
)

# html_descriptions = soup.find_all(
#     'div',
#     class_=r'MuiBox-root jss572 high-risk__HighRiskCard-o7qs9y-0 gdBVup'
# )



high_risk_addresses = [job_elem.text + ', Hong Kong' for job_elem in html_addresses]
# high_risk_descriptions = [job_elem.text for job_elem in html_descriptions]

high_risk_addresses

with open(r'', 'rb') as f:
    coordinates_df = pickle.load(f)

def pop_address(address):
    address_lst = address.split(',')
    return ','.join(address_lst[1:])
    
def get_coordinates(address):
    trial0 = 0
    trial1 = 0
    trial2 = 0
    location = None
    geolocator = Nominatim(user_agent='hk_explorer')
    while location is None and trial0 < 5:
        trial0 += 1
        try:
            location = geolocator.geocode(address)
        except:
            pass
        time.sleep(1)
    if location is None:
        simplified_address1 = pop_address(address)
        while location is None and trial1 < 5:
            trial1 += 1
            try:
                location = geolocator.geocode(simplified_address1)
            except:
                pass
            time.sleep(1)
    if location is None:
        simplified_address2 = pop_address(simplified_address1)
        while location is None and trial2 < 5:
            trial2 += 1
            try:
                location = geolocator.geocode(simplified_address2)
            except:
                pass
            time.sleep(1)
    if location:
        return (location.latitude, location.longitude)
    else:
        return None, None

for address in high_risk_addresses:
    if address in coordinates_df['address'].values:
        pass
    else:
        latitude, longitude = get_coordinates(address)
        coordinates_df = coordinates_df.append(
            pd.Series({
                'loc_id': coordinates_df['loc_id'].max() + 1,
                'address': address,
                'latitude': latitude,
                'longitude': longitude
            }),
            ignore_index=True
        )

coordinates_df

with open(r'', 'wb') as f:
    pickle.dump(coordinates_df,f)

## Simple Visualization using Folium

map_hk = folium.Map(location=[hk_latitude, hk_longitude], zoom_start=9, tiles='stamentoner')
# folium.TileLayer('stamentoner').add_to(map_hk)

for idx, row in coordinates_df.iterrows():
    address = row['address']
    latitude = row['latitude']
    longitude = row['longitude']
#     print(latitude)
    if not np.isnan(latitude) and not np.isnan(longitude):
#         print(f'add {address}')
        folium.CircleMarker(
            [latitude, longitude],
            radius=5,
            popup=address,
            color='red',
            fill=True,
            fill_color='#3186cc',
            fill_opacity=0.7,
            parse_html=False
        ).add_to(map_hk)

map_hk

## Plotly Visualization

with open(r'../.mapbox_token', 'rb') as f:
    token = pickle.load(f)

import plotly.express as px
px.set_mapbox_access_token(token)
fig = px.scatter_mapbox(
    coordinates_df.dropna(), 
    lat="latitude", 
    lon="longitude",     
    hover_name = 'address',
    zoom=10,
    title=r'Coronavirus High Risk Areas',
    size=[1] * coordinates_df.shape[0],
    size_max=6,
    height=900
)
fig.show()
