==> exploration/02_compose_data.ipynb <==
import pandas as pd
import pickle
import numpy as np
import sys
import requests
import json
import pyodbc
import csv
sys.path.insert(1, '../src')

## Hospitals Addresses (this one should be static)

with open(r'../data/HOSPITALS.pkl', 'rb') as f:
    hospitals_df = pickle.load(f)

hospitals_df.head()

# hospitals_df.to_csv(r'../data/HOSPITALS.tsv', index=False, sep='\t')

## High Risk Address

from webscraper import fetch_highrisk

high_risk_df = fetch_highrisk()

# with open(r'../data/HIGH_RISK.pkl', 'wb') as f:
#     pickle.dump(high_risk_df, f)

high_risk_df.to_csv(r'../data/HIGH_RISK.tsv', index=False, sep='\t')

high_risk_df

with open(r'../data/HIGH_RISK.pkl', 'rb') as f:
    high_risk_df = pickle.load(f)

high_risk_df

## Casese

with open(r'../data/CASES.pkl', 'rb') as f:
    cases_df = pickle.load(f)

from webscraper import fetch_cases

cases_df = fetch_cases()

cases_df

# with open(r'../data/CASES.pkl', 'wb') as f:
#     pickle.dump(cases_df, f)

cases_df.replace(r'\n',' ', regex=True).replace(r'\r',' ', regex=True).to_csv(r'../data/CASES.tsv', index=False, sep='\t')

## Awaiting

with open(r'../data/AWAITING.pkl', 'rb') as f:
    awaiting_df = pickle.load(f)

from webscraper import fetch_awaiting_time

awaiting_df = fetch_awaiting_time()

# with open(r'../data/AWAITING.pkl', 'wb') as f:
#     pickle.dump(awaiting_df, f)

# awaiting_df.to_csv(r'../data/AWAITING.tsv', index=False, sep='\t')

## Basic Statistics

with open(r'../data/STATS.pkl', 'rb') as f:
    stats_df = pickle.load(f)

stats_df

from wuhan_functions import calculate_stats

stats_df = calculate_stats(cases_df)

stats_df

with open(r'../data/STATS.pkl', 'wb') as f:
    pickle.dump(stats_df, f)

stats_df.to_csv(r'../data/STATS.csv', index=False)

## Load pickle file

from wuhan_functions import load_data_csv, load_data_pkl

len(load_data_pkl())

len(load_data_csv())

## Load data using sql


server = 'datavisualizationfti.database.windows.net'
database = 'myVisualizationData'
username = 'azureuser'
password = 'Pass7890'
driver = 'SQL Server'  


cnxn = pyodbc.connect('DRIVER={' + driver + '};SERVER='+server+';DATABASE='+database+';UID='+username+';PWD='+ password)


high_risk_df = pd.read_sql('SELECT * FROM HIGH_RISK', cnxn)
cnxn.close()

from wuhan_functions import load_data_sql

cases_df, high_risk_df, stats_df, hospital_awaiting_df = load_data_sql()
