==> exploration/03_understand_data.ipynb <==
import pickle
import numpy as np
import pandas as pd
from pandas_profiling import ProfileReport

## Stats

with open(r'../data/STATS.pkl', 'rb') as f:
    stats_df = pickle.load(f)

stats_df

## Address

with open(r'../data/ADDRESS.pkl', 'rb') as f:
    address_df = pickle.load(f)

address_profile = ProfileReport(address_df, title='Pandas Profiling Report', html={'style':{'full_width':True}})
address_profile

## HIGH_RISK

with open(r'../data/HIGH_RISK.pkl', 'rb') as f:
	high_risk_df = pickle.load(f)

high_risk_df_profile = ProfileReport(high_risk_df, title='Pandas Profiling Report', html={'style':{'full_width':True}})
high_risk_df_profile

## Cases

with open(r'../data/CASES.pkl', 'rb') as f:
	cases_df = pickle.load(f)

cases_df_profile = ProfileReport(cases_df, title='Pandas Profiling Report', html={'style':{'full_width':True}})
cases_df_profile

cases_df
