==> exploration/04_investigate_address_problems.ipynb <==
import pickle
import pandas as pd

address_df_path = r'../data/ADDRESS.pkl'
address_old_df_path = r'../data/ADDRESS_old.pkl'
high_risk_df_path = r'../data/HIGH_RISK.pkl'
high_risk_old_df_path = r'../data/HIGH_RISK_old.pkl'

with open(address_df_path, 'rb') as f:
    address_df = pickle.load(f)
with open(address_old_df_path, 'rb') as f:
    address_old_df = pickle.load(f)
    
with open(high_risk_df_path, 'rb') as f:
    high_risk_df = pickle.load(f)
    
with open(high_risk_old_df_path, 'rb') as f:
    high_risk_old_df = pickle.load(f)

address_df.shape

address_old_df.shape


address_df = address_df.drop_duplicates(subset=['location_en', 'longitude', 'latitude'])

address_df.head()

high_risk_df.head()

high_risk_old_df.head()

pd.merge(
    high_risk_df,
    high_risk_old_df,
    how='left',
    left_on='id',
    right_on='id'
)[['location_zh_x', 'location_zh_y']]

high_risk_with_coordinates_df = pd.merge(
	high_risk_df,
	address_df,
	how='left',
	left_on='location_en',
	right_on='location_en'
)

high_risk_with_coordinates_df

high_risk_with_coordinates_df.shape

import requests

from lxml import html

page = requests.get(r'http://www.kge-reeds.cn/')

tree = html.fromstring(page.content)

tree
