==> exploration/05_For_flourish_visualization.ipynb <==
import os
import sys
sys.path.insert(1, os.path.join('..', 'src'))
import pickle
import numpy as np
import pandas as pd

from wuhan_functions import load_data_live, load_data_pkl, load_data_csv

## Load Live Data

def fetch_stat():
    """
    Return a DataFrame that represent death, confirmed, investigating and reported numbers respectively.
    
    There are two elements in the json that contain the required value: 
    "allBotWarsLatestFigures" and "allWarsLatestFiguresOverride".
    
    The values in allBotWarsLatestFigures are the record in history for every day.
    The values in allWarsLatestFiguresOverride are the live values.
    Whenenver an attribute in allWarsLatestFiguresOverride is not an empty string,
    it overrides the values in allBotWarsLatestFigures.
    
    """
    statnames = ['Death', 'Confirmed', 'Investigating', 'Reported']
    
    page = requests.get('https://wars.vote4.hk/page-data/cases/page-data.json')
    data = json.loads(page.content)
    print(json.dumps(data, indent=4, sort_keys=True))
    stats_yesterday = data['result']['data']['allBotWarsLatestFigures']['edges'][0]
    stats_live = data['result']['data']['allWarsLatestFiguresOverride']['edges'][0]
    
    res = {}
    
    for attr in ['death', 'confirmed', 'investigating', 'reported', 'discharged', 'ruled_out']:
        if stats_live['node'][attr] != "":
            res[attr] = stats_live['node'][attr]
        else:
            res[attr] = stats_yesterday['node'][attr]
    
    df = pd.DataFrame(data=res, index=[0])
    
    return df

fetch_stat()

cases_df, high_risk_df, stats_df, hospital_awaiting_df = load_data_live()

from webscraper import fetch_highrisk

from lxml import html
import requests
import pandas as pd
import numpy as np
import re
from bs4 import BeautifulSoup
import json

df = fetch_highrisk()

flourish_df = df[['lat', 'lng', 'start_date', 'end_date', 'action_en']]

flourish_df = flourish_df[
    (flourish_df['lat'] != '') &
    (flourish_df['lng'] != '') &
    ~(flourish_df['start_date'].isna()) &
    ~(flourish_df['end_date'].isna())
]

flourish_df['start_date'] = flourish_df['start_date'].dt.strftime("%Y-%m-%dT%H:%M:%S.000Z")
flourish_df['end_date'] = flourish_df['end_date'].dt.strftime("%Y-%m-%dT%H:%M:%S.000Z")

flourish_df.shape

flourish_df['Value'] = 1

flourish_df.to_excel('flourish_df.xlsx')

flourish_df['lat'].min()


flourish_df['lat'].max()

flourish_df['lng'].min()

flourish_df['lng'].max()

flourish_df['lng'].sort_values()
